update(scripting): fix type in for

+++ src/components/Plugins/src/Scripting.cpp
@@ -35,7 +35,7 @@ SMM_API void scripting_Print(const char *str)
     std::string message(str);
     if (message.size() != 0)
     {
-        for (int i = 0; i < print_colors.size(); i++)
+        for (size_t i = 0; i < print_colors.size(); i++)
         {
             message = replace(message, print_colors.at(i), "");
             message = replace(message, str_tolower(print_colors.at(i)), "");
